fix(st-09063): correct split test type imports
